Refactor JobRateLimitAlgorithm SPI
